refactor: add menu package
