+++ src/main/java/kitchenpos/menu/domain/MenuRepository.java
@@ -1,8 +1,9 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import kitchenpos.menu.domain.Menu;
 
 public interface MenuRepository {
     Menu save(Menu menu);
